[FIX] Sustituimos GLUTGetTime por clock/CLOCKS_PER_SEC (mucha más precisión)

+++ 2D/Player.h
@@ -18,7 +18,7 @@ class Player
 
 public:
 	void init(const glm::ivec2 &tileMapPos, ShaderProgram &shaderProgram);
-	void update(int deltaTime);
+	void update(float deltaTime);
 	void render();
 
 	void setTileMap(TileMap *tileMap);
